compute mean of arguments - with fix

+++ examples/3B/fix_link_error1/stats.cpp
@@ -2,7 +2,7 @@
 
 #include "stats.h"
 
-double compute_mean (const std::vector<double>& data)
+double compute_mean (const std::vector<float>& data)
 {
   double sum = 0.0;
   for (const auto x : data)
